Reorganize utils and separate graph utils
